backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
pymysql>=1.1.2
gunicorn>=21.2.0
python-dotenv>=1.0.0
markdown>=3.4.0
Pillow>=10.0.0
django-cors-headers>=4.0.0
pypinyin>=0.50.0
cryptography>=42.0.0
whitenoise>=6.6.0
